second commit

+++ p2/scheduler.c
@@ -118,7 +118,7 @@ void schedule(void) {
     }
 
     if (candidate->status == STATUS_) {
-        uint32_t esp = (uint32_t) candidate->stack.memory + SZ_STACK; // Set the stack pointer to the top of the candidate's stack
+        uint64_t esp = (uint64_t) candidate->stack.memory + SZ_STACK; // Set the stack pointer to the top of the candidate's stack
 
 __asm__ volatile (
     "mov %0, %%esp" // Move the esp value to the esp register
